+++ lib/ayeaye/runtime/task_message.py
@@ -37,9 +37,12 @@ class AbstractTaskMessage:
 class TaskPartition(AbstractTaskMessage):
     model_cls: Any  # Not really any but a `Class`; this won't serialise to JSON
     method_name: str
-    method_kwargs: dict
+    method_kwargs: dict = field(default_factory=dict)
     model_construction_kwargs: dict = field(default_factory=dict)
     partition_initialise_kwargs: dict = field(default_factory=dict)
+    # in addition to current resolver context (see :class:`ayeaye.connect_resolve.ConnectorResolver`)
+    # optionally add these when running this task
+    additional_context: dict = field(default_factory=dict)
 
 
 @dataclass